==> 2-data cleaning and preparation.Rmd <==
---
title: "master_greenpat_complexity"
author: "Fabian Scheifele"
date: "29 4 2022"
output: html_document
---


```{r}
suppressMessages(memory.limit(size = NA))

if(!require(install.load)){
  install.packages("install.load")
  library(install.load)
}
suppressMessages(install_load("tidyverse", "dplyr", "ggplot2", "readr", "readxl","knitr","here", "tidyverse" , "anytime" , "data.table", "corrplot", "naniar", "fuzzyjoin","expss", "janitor","readr", "data.table", "labelled", "countrycode"))

`%notin%` <- Negate(`%in%`)
```
1. Import data 
```{r}
y2_raw<-fread(here("data","raw","y02_allpatents.csv"), header=T, fill = T,
             sep= ",", na.strings = c("",NA), data.table = F,
             stringsAsFactors = FALSE)

#Check that there are no complete duplicates
y2_raw<-y2_raw%>%distinct()
#rows stay the same so no complete duplicates in data

#Rename some variables
y2_raw<-y2_raw%>%rename(cpc=cpc_class_symbol)
y2_raw<-y2_raw%>%rename(row_no=V1)

length(unique(y2_raw$person_ctry_code))
#248 countries, a lot? Maybe some old soviet countries need renaming?

length(unique(y2_raw$appln_id))
#4,6 Mio. Different Applications

length(unique(y2_raw$docdb_family_id))
#2.548.400 unique families

length(unique(y2_raw$appln_id))/length(unique(y2_raw$docdb_family_id))
#1.8 applications per family

length(unique(y2_raw$cpc))
#476 different CPC codes

length(unique(y2_raw$earliest_filing_id))
#2.493,459 different priorities 

length(unique(y2_raw$appln_auth))
#97 different patent offices

sum(is.na(y2_raw$person_ctry_code))/length(y2_raw$person_ctry_code)
#Residence country of inventor unknown for 47.8% of the data. 8.6 million entries


```

2. Clean data
```{r}
#remove spaces in CPC codes for matching with descriptions
y2_raw<-y2_raw%>%mutate(cpc=str_replace(cpc," ",""))

#make all string variables to lower case
y2_raw <- y2_raw %>%
  mutate(across(where(is.character), tolower))


#Code factor variable granted
y2_raw$granted<- ifelse(y2_raw$granted=='y',1,0)
summary(y2_raw$granted)

```

3. Filter for post-1995 period
```{r}
y2_final95<-y2_raw%>%filter(earliest_filing_year>=1995)
remove(y2_raw)
```

4. Replace names of countries that do no longer exists
```{r}
old_countr<-y2_final95%>%filter(appln_auth %in% c("su","dd","cs","yu","yd","sy"))
length(unique(old_countr$appln_auth))
#only YU was continued after 1995-2006, but would replace with Serbia to have better time series. As filing under Montenegro code starts in 1997, but RS only 2003, it can be assumed that remaining YU mostly serbian (also because of country size)

y2_final95["appln_auth"][y2_final95["appln_auth"] == "yu"] <- "rs"
length(unique(y2_final95$appln_auth))

test<-y2_final95%>%filter(person_ctry_code %in% c("su","dd","cs","yu"))
length(unique(test$person_ctry_code))

#Replacing country codes for "yu" "cs" "su" "dd"
#While not all of the Soviet Union was Russian, it was the largest nation hence SU to RU (808 observations), same for Czech Republic and Czech-slovakia (56 observations)
y2_final95["person_ctry_code"][y2_final95["person_ctry_code"] == "dd"] <- "de"
y2_final95["person_ctry_code"][y2_final95["person_ctry_code"] == "yu"] <- "rs"
y2_final95["person_ctry_code"][y2_final95["person_ctry_code"] == "cs"] <- "cz"
y2_final95["person_ctry_code"][y2_final95["person_ctry_code"] == "su"] <- "ru"
y2_final95["person_ctry_code"][y2_final95["person_ctry_code"] == "uk"] <- "gb"

#make country codes upper case
y2_final95$appln_auth<-toupper(y2_final95$appln_auth)
y2_final95$person_ctry_code<-toupper(y2_final95$person_ctry_code)
y2_final95$cpc<-toupper(y2_final95$cpc)

remove(old_countr)
remove(test)
```

5. removing remaining Duplicate combinations of Application ID, person ID and CPC
```{r}
y2_final95_distinct<-y2_final95%>%
  distinct(appln_id, person_id, cpc, .keep_all=TRUE)

#3000 entries seem to have duplicate IPC, Person_ID and appln_id, but are not complete duplicates. Could be additional documents relating to the same application? 
remove(y2_final95)
```

6. Replace deleted CPC codes by latest revisions
```{r}
cpc_changes<-read_xlsx(here("data","raw","cpc_changes.xlsx"))

#remove white space
cpc_changes$old_cpc<-str_replace(cpc_changes$old_cpc," ","")
cpc_changes$new_cpc<-str_replace(cpc_changes$new_cpc," ","")
y2_final95_distinct$cpc<-str_replace(y2_final95_distinct$cpc," ","")

#merge with data
y2_final95_distinct<-y2_final95_distinct%>%
  left_join(cpc_changes, by=c("cpc"="old_cpc"))

sum(is.na(y2_final95_distinct$new_cpc))/length(y2_final95_distinct$appln_id)
#less than 0.05% of observations are affected by this

#replace codes that have been replaced by other codes
y2_final95_distinct<-y2_final95_distinct%>%
 mutate(cpc_updated=case_when(is.na(new_cpc)~cpc,TRUE~new_cpc))

sum(is.na(y2_final95_distinct$cpc_updated))
#replacement worked, not NA's anymore#

#Replace the remaining 13 codes that were filed but no longer exist manually
y2_final95_distinct$cpc_updated[y2_final95_distinct$cpc_updated=="Y02D10/13"
 | y2_final95_distinct$cpc_updated=="Y02D10/13"
| y2_final95_distinct$cpc_updated=="Y02D10/126"
| y2_final95_distinct$cpc_updated=="Y02D10/171"
| y2_final95_distinct$cpc_updated=="Y02D10/172"
| y2_final95_distinct$cpc_updated=="Y02D10/173"
| y2_final95_distinct$cpc_updated=="Y02D10/122"
| y2_final95_distinct$cpc_updated=="Y02D10/16"
| y2_final95_distinct$cpc_updated=="Y02D10/14"
| y2_final95_distinct$cpc_updated=="Y02D10/151"
| y2_final95_distinct$cpc_updated=="Y02D10/153"]<-"Y02D10/00"

y2_final95_distinct$cpc_updated[y2_final95_distinct$cpc_updated=="Y02D70/00"]<-"Y02D30/70"

#B60/126 was actually moved to Y02D/157 but then this was again replaced by Y02D10/00
y2_final95_distinct$cpc_updated[y2_final95_distinct$cpc_updated=="Y02B60/126"]<-"Y02D10/00"

y2_final95_distinct$cpc_updated[y2_final95_distinct$cpc_updated=="Y02T50/14"]<-"Y02T50/10"

#Check with latest description file to see any missings (i.e. subgroups that do not exist)
cpc_desc<-read_xlsx(here("data","raw","cpc_description.xlsx"))
cpc_desc$CPC<-str_replace(cpc_desc$CPC," ","")
y2_final95_distinct<-y2_final95_distinct%>%
  left_join(cpc_desc, by=c("cpc_updated"="CPC"))

cpc_notfound<-y2_final95_distinct%>%filter(is.na(description))%>%distinct(cpc_updated)


#dummy to mark them
y2_final95_distinct<-y2_final95_distinct%>%
 mutate(cpc_replaced=case_when(cpc==cpc_updated~0,TRUE~1))

#remove text columns
y2_final95_distinct<-y2_final95_distinct%>%
  select(-c("description","level"))

#removing any final duplicates of CPCs that may have been created through CPC renaming
y2_final95_distinct<-y2_final95_distinct%>%
  distinct(appln_id, person_id, cpc, .keep_all=TRUE)
```


6. Reduce to international patent families
```{r}
#Using IPF methodology from IEA (2021): Patents and the energy transition 
#https://iea.blob.core.windows.net/assets/b327e6b8-9e5e-451d-b6f4-cbba6b1d90d8/Patents_and_the_energy_transition.pdf

#1select international and regional patents
intreg_pat<-y2_final95_distinct%>%
  filter(appln_auth %in% c("EP", "WO", "AP", "EM","BX", "EA", "GC","EU", "IB", "OA", "QZ", "XN", "XU", "XV"))

#2reduce from applications to families
intreg_families<- intreg_pat %>%
  distinct(docdb_family_id, person_id, cpc, .keep_all = TRUE)
#create different office variable to be able to append with national families
intreg_families$off_dist<-0

#3Select nationally filed patents#
nat_pat<-y2_final95_distinct%>%
  filter(appln_auth %notin% c("EP", "WO", "AP", "EM","BX", "EA", "GC","EU", "IB", "OA", "QZ", "XN", "XU", "XV"))

#4Create variable to see in how many different offices national patents were filed#
nat_pat <-nat_pat %>%
  group_by(docdb_family_id) %>%
  mutate(off_dist = n_distinct(appln_auth))

summary(nat_pat$off_dist)
hist(nat_pat$off_dist)

#5 Keep only national families filed in more than country
national_families <- nat_pat %>%
  distinct(docdb_family_id, person_id, cpc, .keep_all = TRUE) %>%
  filter(off_dist > 1) 
summary(national_families$off_dist)

#6 append international and national families with 2+offices back together
int_pat_fam<-bind_rows(national_families,intreg_families)

#7 filter out any remaining family duplicates that may arise out of a family that is both inter/regional as well as filed in 2+ national offices
int_pat_fam_final<-int_pat_fam %>%
  distinct(docdb_family_id, person_id, cpc, .keep_all = TRUE)

sum(is.na(int_pat_fam_final$person_ctry_code))/length(int_pat_fam_final$person_ctry_code)
#still 28% without person_ctry_code, compared to 47% originally

```

7. Checking out missing country information for IPF's
```{r}
missing_country<-int_pat_fam_final%>%
  filter(is.na(person_ctry_code))

length(unique(missing_country$docdb_family_id))/length(unique(int_pat_fam_final$docdb_family_id))
#For 58.4% of all IPF inventor residence info is incomplete. 

family_found<-int_pat_fam_final%>%filter(!is.na(person_ctry_code))%>%
  distinct(docdb_family_id)

length(family_found$docdb_family_id)/length(unique(int_pat_fam_final$docdb_family_id))
#For 98.4% of International Patent Families we have at least one inventor's residence info

missing_country%>%group_by(appln_auth)%>%
  summarise(top_missing=n_distinct(docdb_family_id))%>%
  slice_max(top_missing, n=10)
#Patent offices with most missing info are CN, JP, AU, KR, ES, MX
```

8. Attempt to recuperate remaining missing data through matching with other applications
that were dropped by distinct: Nothing found! 
```{r}

#match with original raw data
#country_exist<-y2_final95_distinct%>%filter(!is.na(person_ctry_code))
#country_exist<-country_exist%>%select(appln_id,docdb_family_id,person_id,person_ctry_code)
#filter unique combinations
#countries_exist_dis<-country_exist%>%distinct(person_id, person_ctry_code)

#countries_found<-left_join(missing_country,countries_exist_dis, by=c("person_id"))
#Does not find any missing person information
#sum(is.na(countries_found$person_ctry_code.y))
```

9. Dropping remaining NA's for person_ctry_code and creating final data set
```{r}
#Dropping obsevations without country information
int_pat_fam_final_complete<-int_pat_fam_final%>%filter(!is.na(person_ctry_code))

int_pat_fam_final%>%group_by(appln_auth)%>%
  summarise(count=n_distinct(docdb_family_id))%>%arrange(desc(count))

int_pat_fam_final_complete%>%
  group_by(appln_auth)%>%
  summarise(count=n_distinct(docdb_family_id))%>%arrange(desc(count))
#Japan has almost zero origin data, China families also reduced by 90%

int_pat_fam_final%>%filter(is.na(person_ctry_code))%>%
  group_by(appln_auth)%>%
  summarise(count=n_distinct(docdb_family_id))%>%arrange(desc(count))


#write final family-level data set (unique family, person, ipc combination)
write_rds(int_pat_fam_final_complete, here("data", "intermediate","Y02_intl_pat_families_1995.rds"))


#IF WANTED FOR ROBUSTNESS ONE COULD replace npw the missing country values in int_pat_fam_final with the application authority of the priority patent! 
```


10. Seperating into two datasets prior to counting (international families and all patents)
```{r}
#Final dataset: 14,7 mio entries, from originally 18.155.054 however 42% replaced nationality values

#saveRDS(y2_final95_distinct,here
       # ("data","intermediate","y2_patentlevel_final_1995.rds"))

#Create final application-level set without person_country_code replacements
y2_final_norepl<-y2_final95_distinct%>%filter(!is.na(person_ctry_code))
saveRDS(y2_final_norepl,here
        ("data","intermediate","y2_patentlevel_final_1995_noreplacements.rds"))

```
