==> 3-count data_creation_descriptives.Rmd <==
---
title: "Creating count data and descriptives"
author: "Fabian Scheifele"
date: "3 6 2022"
output: html_document
uses: "y2_final95.rds"
creates: "y2_count.rds"
editor_options: 
  chunk_output_type: console
---

1. packages
```{r}
suppressMessages(memory.limit(size = NA))

if(!require(install.load)){
  install.packages("install.load")
  library(install.load)
}
suppressMessages(install_load("tidyverse", "dplyr", "ggplot2", "readr", "readxl","knitr","here", "tidyverse" , "anytime" , "data.table", "corrplot", "naniar", "fuzzyjoin","expss", "janitor","readr", "data.table", "labelled", "countrycode","grep", "economiccomplexity"))

`%notin%` <- Negate(`%in%`)
```

2. import data
```{r}
y2_fam_final<-readRDS(here("data","intermediate","Y02_intl_pat_families_1995.rds"))
#rename updated cpc for code cohesion and drop old cpc variable
y2_fam_final<-y2_fam_final%>%
  rename(old_cpc=cpc)
y2_fam_final<-y2_fam_final%>%
  rename(cpc=cpc_updated)

#how many distinct IPC codes per firm
y2_fam_final<-y2_fam_final%>%group_by(docdb_family_id)%>%
  mutate(distinct_cpc=n_distinct(cpc), distinct_inv=n_distinct(person_id), distinct_year=n_distinct(earliest_filing_year))

summary(y2_fam_final$distinct_cpc)
summary(y2_fam_final$distinct_inv)
summary(y2_fam_final$distinct_year)

different_years<-y2_fam_final%>%filter(distinct_year>1)
length(unique(different_years$docdb_family_id))
#For 2459 patents the earliest_filing_year is not uniform across different applications of the same family, in those cases the earliest year is taken as family filing year#
y2_fam_final<-y2_fam_final%>%
  group_by(docdb_family_id)%>%
  mutate(earliest_fam_filing_year=min(earliest_filing_year))
  
length(unique(y2_fam_final$earliest_fam_filing_year))/length(unique(y2_fam_final$earliest_fam_filing_year))
#no filing year unique

y2_fam_final<-y2_fam_final%>%arrange(desc(distinct_inv), docdb_family_id)

large_families<-y2_fam_final%>%filter(distinct_inv>40)
length(unique(large_families$docdb_family_id))

length(unique(large_families$docdb_family_id))/length(unique(y2_fam_final$docdb_family_id))

inv_counts<-y2_fam_final%>%distinct(docdb_family_id,distinct_inv)
summary(inv_counts$distinct_inv)
hist(inv_counts$distinct_inv)


```


3. Creating fractional count by appln_id for y2_final (with replaced residences)
```{r}
#Create fractional count (each family-person-CPC combination is one row with the fraction of this inventor-CPC combination)
y2_fam_final<-y2_fam_final%>%
  group_by(docdb_family_id)%>%
  mutate(frac=1/n())

#test to show that person-cpc combination is indeed one row
test<-y2_fam_final%>%filter(appln_id==488313783)%>%select(docdb_family_id, appln_id, person_id,cpc)%>%
  arrange(person_id,cpc)

#Calculate Patent Counts
pat_count<-y2_fam_final%>%
  group_by(person_ctry_code, earliest_fam_filing_year, cpc)%>%
  summarize(pat_count=sum(frac))

#remove white space in between
pat_count$cpc<-gsub(" ", "", pat_count$cpc, fixed = TRUE)

#Merge CPC descriptions to data#
cpc_test<-read_excel(here("data","raw","cpc_description.xlsx"), col_names = TRUE)
cpc_test$CPC<-gsub(" ", "", cpc_test$CPC, fixed = TRUE)
pat_count<-left_join(pat_count, cpc_test, by=c("cpc"="CPC"))

#merge countrycodes to data
#country_codes<-read_excel(here("data","wipo_codes.xlsx"))
library(countrycode)
pat_count$country<-countrycode(pat_count$person_ctry_code, origin = 'iso2c', destination = 'country.name')

nonames<-pat_count%>%
  filter(is.na(country))
#80 observations either relate to outdated country codes or to accidently put patent office codes, 
#observations are dropped as they cannot be attributed to specific country

pat_count<-pat_count%>%filter(!is.na(country))
```

4.Calculate (unrestricted) Revealed Technological advantages
```{r}
pat_count<-pat_count%>%group_by(country,earliest_fam_filing_year)%>%
  mutate(sum_country_year=sum(pat_count))%>%
  ungroup()%>%
  group_by(cpc, earliest_fam_filing_year)%>%
  mutate(sum_cpc_year=sum(pat_count))%>%
  ungroup()%>%
  group_by(earliest_fam_filing_year)%>%mutate(total_pat_year=sum(pat_count))%>%
  ungroup()%>%
  mutate(rta=(pat_count/sum_country_year)/(sum_cpc_year/total_pat_year))
```
5. rename for easier use
```{r}
pat_count<-rename(pat_count,year=earliest_fam_filing_year)
```



5. Some tests on the data
```{r}
pat_count%>%slice_max(rta,n=20)%>%select(country, year, cpc, rta)

country_year<-pat_count%>%group_by(country, year)%>%
  summarise(count=sum(pat_count))

#Yearly count shows that patents are dropping in 2020, 2021, so exclude
pat_count%>%
  group_by(year)%>%
  summarise(count=sum(pat_count))%>%
  ggplot(aes(x=year,y=count))+geom_line()

top_countries<-pat_count%>%
  group_by(country)%>%
  summarise(count=sum(pat_count))%>%
  slice_max(count,n=15)

#plot with top 15 countries
ggplot(data=top_countries, aes(x=country,y=count))+
  geom_bar(stat="identity")+ggtitle("15 Countries with most Y02 Patent Families 1995-2021")+geom_text(aes(label=round(count,digits=0)),vjust=-0.4)

#plot with main CPC classes
pat_count$description<-as.factor(pat_count$description)
levels(pat_count$description) <- gsub(" ", "\n", levels(pat_count$description))
pat_count%>%
  group_by(description)%>%
  summarise(count=sum(pat_count))%>% 
  slice_max(count,n=10)%>%
  ggplot(aes(x=description,y=count))+
  geom_bar(stat="identity")+ggtitle("IPC codes with most Y02 Patent Families 1995-2021")+geom_text(aes(label=round(count,digits=0)),vjust=-0.4)


topcpc<-pat_count%>%
  group_by(cpc)%>%
  summarise(count=sum(pat_count))

#CPC codes skewed, some codes with high counts, majority low
topcpc%>%filter(count<5000)%>%
  ggplot(aes(x=count))+geom_histogram()

summary(topcpc$count)
#with non-updated CPCs: median was 131 families per CPC (over the whole period)
#with newly updated CPC: median is 299 families per CPC (over the whole period)

length(unique(pat_count$person_ctry_code))
#205 countries

length(unique(pat_count$cpc))
#with non-updated CPCs: 432 different Y02 cpc tags
#with newly updated CPC:331 different Y02 cpc tags--> there was some consolidation, good!
```


7. Filter out 2020/2021 and countries infrequently filed
```{r}
#First, create balanced panel with all cpc, year, country combinations
pat_count_balanced<-pat_count%>%
  expand(year,cpc,country)

pat_count_balanced_filled<-pat_count%>%right_join(pat_count_balanced)

#Delete patent families in 2020 and 2021 due to strong drop due to time of registration
pat_count_balanced_filled<-pat_count_balanced_filled%>%
  filter(year<2020)

cpcs<-pat_count_balanced_filled%>%
  group_by(cpc,year)%>%
  summarise(family_count=sum(pat_count,na.rm = TRUE))%>%
  filter(family_count==0)

cpcs%>%
  group_by(year)%>%
  summarise(no_cpc_missing=n_distinct(cpc))%>%
  ggplot(aes(x=year, y=no_cpc_missing))+
  geom_bar(stat = "identity")+ggtitle("Number of CPC codes where no single family was filed per year")
  

length(unique(cpcs$cpc))
length(unique(pat_count_balanced_filled$cpc))
#old CPC: for 212 out of 432 CPC codes have at least one year where no family is filled
#new CPC: 114 out of 331 have at least one year where no family is filled#


#Countries with zero family applications in a particular year
countries<-pat_count_balanced_filled%>%
  group_by(country,year)%>%
  summarise(country_count=sum(pat_count,na.rm = TRUE))%>%
  filter(country_count==0)


#Vector of countries that have at least one year with zero intl. family applications#
countries_drop<-unique(countries$country)

countries%>%
  group_by(year)%>%
  summarise(countries_year_miss=n_distinct(country))%>%
  ggplot(aes(x=year, y=countries_year_miss))+
  geom_bar(stat = "identity")+ggtitle("Number of Countries that have no Family application per year")
  
length(unique(countries$country))
length(unique(pat_count_balanced_filled$country))
#157 out of 205 countries have at least one year where no family is filed

length(unique(pat_count_balanced_filled$country))-length(unique(countries$country))
#48 countries will remain in the dataset

#Final set of countries
pat_count_country_final<-pat_count_balanced_filled%>%
  filter(country %notin% countries_drop)

country_names<-pat_count_country_final%>%select(1,7)%>%distinct()

#Add hong-kong to mainland China
pat_count_country_final$
  person_ctry_code[pat_count_country_final$
                                           person_ctry_code=="HK"]<-"CN"

pat_count_country_final$
  country[pat_count_country_final$country=="Hong Kong SAR China"]<-"China"

#Take out small city states (Liechtenstein & Luxemburg)

pat_count_country_final<-pat_count_country_final%>%
  filter(country %notin% c("Liechtenstein", "Luxembourg"))

#remove unused cpc-country combinations created by "expand"
pat_count_country_final<-pat_count_country_final%>%
  filter(!is.na(pat_count))

length(unique(pat_count_country_final$country))
#45 countries remain after HK-->CN, Liechtenstein and Luxembourg removed
```

8. Check out infrequent CPCs to identify filter or aggregation method
```{r}
#prepare aggregated level 2 data
level1_2<-cpc_test%>%filter(level %in% c(0,1,2))%>%select(1,2)

cpc_test<-cpc_test%>%left_join(level1_2, by="CPC")

cpc_test<-cpc_test%>%
  mutate(cpc_level12=case_when(is.na(level.y)~NA_character_,TRUE~CPC))

#Fill level 3 and 4 values with nearest above level 2 value
cpc_test<-cpc_test%>%fill(cpc_level12)%>%select(1,5)

#merge to country dataset
pat_count_country_final<-pat_count_country_final%>%
  left_join(cpc_test, by=c("cpc"="CPC"))

#Check for which codes the merge did not work
NAs<-pat_count_country_final%>%filter(is.na(cpc_level12))
#no NA's means that aggregate level was found for all CPC codes

length(unique(pat_count_country_final$cpc))
length(unique(pat_count_country_final$cpc_level12))
#This aggregation would reduce the number of individual patent classes from 190 to 320

#Re-doing the graph from above with the CPC without single application by year

pat_count_cpc_agg<-pat_count_country_final%>%
  group_by(year,cpc_level12)%>%
  summarise(pat_count2=sum(pat_count, na.rm = TRUE))


pat_count_cpc_agg%>%filter(pat_count2<10)%>%
  ggplot(aes(x=pat_count2))+geom_histogram()

summary(pat_count_cpc_agg$pat_count2)
#there is no class with zero counts any more, but still quite some observations were only one application was filed per year (seen through fractional counts less than 1 or 1) 

#Let's aggregate over the whole time period to see which Codes are really rare overall
pat_count_cpc_total<-pat_count_country_final%>%
  group_by(cpc_level12)%>%
  summarise(pat_count3=sum(pat_count))

pat_count_cpc_total%>%filter(pat_count3<1000)%>%
  ggplot(aes(x=pat_count3))+geom_histogram()

summary(pat_count_cpc_total$pat_count3)

#Possible threshold of very infrequent CPC classes

# 10 or less counts over the whole period: eliminates 16/190 classes
# less than 10 counts over the whole period: eliminate 15/190 classes
#less than 25 counts over the whole period: eliminates 22/190 classes
#last one would make most sense because we observe 24 years of data so then on 
#average each class has at least one count per year (is of course not equally distributed)

y02<-read_excel(here("data","raw","cpc_description.xlsx"), col_names = TRUE)
small_cpcs<-pat_count_cpc_total%>%
  filter(pat_count3<25)%>%left_join(y02, by=c("cpc_level12"="CPC"))

cpc_drop<-small_cpcs$cpc_level12
#15 out the 22 categories we are loosing are high level-1 categories, which usually not filed as patents aim for more specific patents codes (levels 2-4). The remainin 7 relate to agro-forestry, chemicals and the Y2B10 actually only refers to CCS in the context building technologies, normal CCS has separate category Y2C. 

#Patents are of course not equally distributed across years.
pat_count_cpc_agg%>%group_by(year)%>%
  summarise(pat_py=sum(pat_count2))%>%
  ggplot(aes(x=year,y=pat_py))+
  geom_bar(stat="identity")+ggtitle("Patent Family counts per year in aggregated categories")+geom_text(aes(label=round(pat_py,digits=0)),vjust=-0.4, size=2.5)


```

9. Final CPC selection: Filter out CPC<25 counts over full period and use level-2 aggregated data
```{r}
pat_count_country_reduced<-pat_count_country_final%>%
  group_by(year,country, cpc_level12)%>%
  summarise(ipf_count=sum(pat_count))

#filter out the identified cpc codes#
pat_count_country_reduced<-pat_count_country_reduced%>%
  filter(cpc_level12 %notin% cpc_drop)

#Re-Do RTA calculations with reduced and aggregated set
pat_count_country_reduced<-pat_count_country_reduced%>%
  group_by(country,year)%>%
  mutate(sum_country_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(cpc_level12, year)%>%
  mutate(sum_cpc_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(year)%>%mutate(total_pat_year=sum(ipf_count))%>%
  ungroup()%>%
  mutate(rta=(ipf_count/sum_country_year)/(sum_cpc_year/total_pat_year))

#rename country variable so it works with the economiccomplexity package
pat_count_country_reduced<-pat_count_country_reduced%>%
  rename(origin=country)

#write final rds
saveRDS(pat_count_country_reduced, here("data","final","final_ipf_count_reduced.rds"))
```

10. Descriptive Statistics on final count data set
```{r}

```
