==> 4- complexity calculations.Rmd <==
---
title: "4- Complexity calculations"
author: "Fabian Scheifele"
date: "5 7 2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

1. packages
```{r}
suppressMessages(memory.limit(size = NA))

if(!require(install.load)){
  install.packages("install.load")
  library(install.load)
}
suppressMessages(install_load("tidyverse", "dplyr", "ggplot2", "readr", "readxl","knitr","here", "tidyverse" , "anytime" , "data.table", "corrplot", "naniar", "fuzzyjoin","expss", "janitor","readr", "data.table", "labelled", "countrycode","grep", "economiccomplexity","xtable"))

`%notin%` <- Negate(`%in%`)

rta<-function(df,i,t,cat,value){
  df%>%group_by(i,t)%>%
  mutate(sum_country_year=sum(value))%>%
  ungroup()%>%
  group_by(cat, t)%>%
  mutate(sum_cpc_year=sum(value))%>%
  ungroup()%>%
  group_by(t)%>%mutate(total_pat_year=sum(value))%>%
  ungroup()%>%
  mutate(rta=(value/sum_country_year)/(sum_cpc_year/total_pat_year))
}
options(scipen=999)
```

2. import data
```{r}
pat_count_country_reduced<-readRDS(here("data","final","final_ipf_count_reduced.rds"))
cpc_text<-read_xlsx(here("data","raw","cpc_description.xlsx"))
```

3. Create general ECI function (to be used for different time frames)
```{r}
#PCI Calculation via function
pci_function <- function(t,df) {
  d_year<- df %>% filter(year == t)
  balassa_mat<-balassa_index(
        data = d_year,
        country = "origin",
        product = "cpc_level12",
        value = "ipf_count"
      )
  com<-complexity_measures(balassa_mat,
                      method = "eigenvalues")
  pci<-as.data.frame(com$complexity_index_product)
  pci<-setDT(pci,keep.rownames = TRUE)%>%mutate(year= t)
  pci<-pci%>%rename(cpc=rn)
  pci<-pci%>%rename(pci="com$complexity_index_product")
}

#ECI Calculation via function
eci_function <- function(t,df) {
  d_year<- df %>% filter(year == t)
  balassa_mat<-balassa_index(
        data = d_year,
        country = "origin",
        product = "cpc_level12",
        value = "ipf_count"
      )
  com<-complexity_measures(balassa_mat,
                      method = "eigenvalues")
  eci<-as.data.frame(com$complexity_index_country)
  eci<-setDT(eci,keep.rownames = TRUE)%>%mutate(year= t)
  eci<-eci%>%rename(country=rn)
  eci<-eci%>%rename(eci="com$complexity_index_country")
}

```


3. Create annual complexity indices (may be a bit too volatile)
```{r}

years<-c(1995:2019)
pciall<-lapply(years, pci_function, 
               df= pat_count_country_reduced)
pciall<-bind_rows(pciall)

eciall<-lapply(years,eci_function, df =pat_count_country_reduced)
eciall<-bind_rows(eciall)

remove(eciall)
remove(pciall)
```


4. Three year intervals
```{r}
pat_count_country_reduced$threeyear_period <-with(pat_count_country_reduced,
       ifelse(year %in% c(1995:1997) , "1995-1997",
              ifelse(year %in% c(1998:2000) , "1998-2000",
              ifelse(year %in% c(2001:2003) , "2001-2003",       
              ifelse(year %in% c(2004:2006) , "2004-2006",       
              ifelse(year %in% c(2007:2009) , "2007-2009", 
              ifelse(year %in% c(2010:2012) , "2010-2012", 
              ifelse(year %in% c(2013:2015) , "2013-2015",
              ifelse(year %in% c(2016:2018) , "2016-2018",
                     NA)))))))))

threeyear<-pat_count_country_reduced%>%filter(!is.na(threeyear_period))%>%
  group_by(threeyear_period,origin, cpc_level12)%>%
  summarise(ipf_count=sum(ipf_count))

threeyear<-threeyear%>%rename(year=threeyear_period)

#replicate RCA calculations for three-year interval
threeyear<-threeyear%>%
  group_by(origin,year)%>%
  mutate(sum_country_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(cpc_level12, year)%>%
  mutate(sum_cpc_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(year)%>%mutate(total_pat_year=sum(ipf_count))%>%
  ungroup()%>%
  mutate(rta=(ipf_count/sum_country_year)/(sum_cpc_year/total_pat_year))

#Replicate ECI and PCI Calculations
years_3y<-unique(threeyear$year)
pci_3y<-lapply(years_3y, pci_function, df=threeyear)
pci_3y<-bind_rows(pci_3y)

eci_3y<-lapply(years_3y, eci_function, df=threeyear)
eci_3y<-bind_rows(eci_3y)

#Join values to main dataset
threeyear<-threeyear%>%left_join(pci_3y, by=c("year","cpc_level12"="cpc"))
threeyear<-threeyear%>%left_join(eci_3y, by=c("year","origin"="country"))
```


5. Comparison of ECI/PCI with annual or three year weighted RCA
```{r}
top_15<-function(t,df){
  table<-df%>%filter(year==t)%>%
  slice_max(eci, n=15)%>%select(1,2)
}

top15<-lapply(years_3y, top_15, df=eci_3y)

lead9597<-eci_3y%>%filter(year=="1995-1997")%>%
  slice_max(eci, n=15)%>%select(1,2)
lead9800<-eci_3y%>%filter(year=="1998-2000")%>%
  slice_max(eci, n=15)%>%select(1,2)
lead0103<-eci_3y%>%filter(year=="2001-2003")%>%
  slice_max(eci, n=15)%>%select(1,2)
lead0406<-eci_3y%>%filter(year=="2004-2006")%>%
  slice_max(eci, n=15)%>%select(1,2)
lead0709<-eci_3y%>%filter(year=="2007-2009")%>%
  slice_max(eci, n=45)%>%select(1,2)
lead1012<-eci_3y%>%filter(year=="2010-2012")%>%
  slice_max(eci, n=45)%>%select(1,2)
lead1315<-eci_3y%>%filter(year=="2013-2015")%>%
  slice_max(eci, n=15)%>%select(1,2)
lead1618<-eci_3y%>%filter(year=="2016-2018")%>%
  slice_max(eci, n=15)%>%select(1,2)

print(xtable(lead9597, caption="ECI for Period 1995-1997 based on Y02 Intl. Patent Families", type = "latex"), file = here("output","eci9597.tex"))
print(xtable(lead0406, caption="ECI for Period 2004-2006 based on Y02 Intl. Patent Families", type = "latex"), file = here("output","eci0406.tex"))
print(xtable(lead0709, caption="ECI for Period 2007-2009 based on Y02 Intl. Patent Families", type = "latex"), file = here("output","eci0709.tex"))
print(xtable(lead1012, caption="ECI for Period 2010-2012 based on Y02 Intl. Patent Families", type = "latex"), file = here("output","eci1012.tex"))
print(xtable(lead1315, caption="ECI for Period 2013-2015 based on Y02 Intl. Patent Families", type = "latex"), file = here("output","eci1315.tex"))

#Create differences of RTA over time
threeyear<-threeyear%>%mutate(rta1=if_else(rta>1,1,0))
```


8. Some info about the 3-year grouped data
```{r}
country_agg<-threeyear%>%group_by(origin,year)%>%
  summarise(sum_ipf=sum(ipf_count), avg_rta=mean(rta), avg_pci=mean(pci), eci=mean(eci), distinct_cpc=n_distinct(cpc_level12), sum_rta=sum(rta1))

test1995<-country_agg%>%filter(year=="1995-1997")
mean(test1995$avg_rta)


threeyear$year<-as.factor(threeyear$year)

year_agg<-threeyear%>%
  group_by(year)%>%
  summarise(sum_ipf=sum(ipf_count), avg_rta=mean(rta), avg_pci=mean(pci), eci=mean(eci), distinct_cpc=n_distinct(cpc_level12), sum_rta=sum(rta1), distinct_countries=n_distinct(origin))


avg_test<-threeyear%>%filter(year=="1995-1997")%>%select(rta)
summary(avg_test$rta)

threeyear%>%filter(rta<20)%>%ggplot(aes(rta))+geom_histogram()+ggtitle("RTA Distribution, cut off at 20")
argentina95<-threeyear%>%filter(origin=="Argentina", year=="1995-1997")
Y02A40_10<-threeyear%>%filter(cpc_level12=="Y02A40/10", year=="1995-1997")
sum(Y02A40_10$ipf_count)

germany<-pat_count_country_reduced%>%filter(origin=="Germany", year %in% c(1995:1997), cpc_level12=="Y02A40/10")

germany2<-threeyear%>%
  filter(origin=="Germany", year =="1995-1997", cpc_level12=="Y02A40/10")

median(threeyear$rta)
```


8. 3-year intervals: What happend with the industrialised countries
```{r}
what_happened<-country_agg%>%
  filter(origin%in%c("Germany", "United States", "China", "South Korea", "Taiwan", "Sweden"))
ggplot(what_happened, aes(x=year,y=avg_rta, color=origin, group=origin))+geom_point()+geom_line()

ggplot(what_happened, aes(x=year,y=avg_pci, color=origin, group=origin))+geom_point()+geom_line()

ggplot(what_happened, aes(x=year,y=sum_ipf, color=origin, group=origin))+geom_point()+geom_line()

ggplot(what_happened, aes(x=year,y=eci, color=origin, group=origin))+geom_point()+geom_line()

ggplot(what_happened, aes(x=year,y=distinct_cpc, color=origin, group=origin))+geom_point()+geom_line()

ggplot(what_happened, aes(x=year,y=sum_rta, color=origin, group=origin))+geom_point()+geom_line()

threeyear%>%group_by(year)%>%
  summarise(ipf_total=sum(ipf_count))%>%
  ggplot(aes(x=year,y=ipf_total))+
  geom_bar(stat="identity")+
  ggtitle("Y02 Family counts by 3-year period")+geom_text(aes(label=round(ipf_total,digits=0)),vjust=-0.4)

```


11. Five year intervals ECI and PCI calculations
```{r}
pat_count_country_reduced$fiveyear_period <-with(pat_count_country_reduced,
       ifelse(year %in% c(1995:1999) , "1995-1999",
              ifelse(year %in% c(2000:2004) , "2000-2004",
              ifelse(year %in% c(2005:2009) , "2005-2009",       
              ifelse(year %in% c(2010:2014) , "2010-2014",       
              ifelse(year %in% c(2015:2019) , "2015-2019", 
                     NA))))))

fiveyear<-pat_count_country_reduced%>%
  group_by(fiveyear_period,origin, cpc_level12)%>%
  summarise(ipf_count=sum(ipf_count))

fiveyear<-fiveyear%>%rename(year=fiveyear_period)

#replicate RCA calculations for five-year interval
fiveyear<-fiveyear%>%
  group_by(origin,year)%>%
  mutate(sum_country_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(cpc_level12, year)%>%
  mutate(sum_cpc_year=sum(ipf_count))%>%
  ungroup()%>%
  group_by(year)%>%mutate(total_pat_year=sum(ipf_count))%>%
  ungroup()%>%
  mutate(rta=(ipf_count/sum_country_year)/(sum_cpc_year/total_pat_year))%>%
  mutate(rta1=if_else(rta>1,1,0))


#Replicate ECI and PCI Calculations
years_5y<-unique(fiveyear$year)
pci_5y<-lapply(years_5y, pci_function, df=fiveyear)
pci_5y<-bind_rows(pci_5y)

fiveyear<-lapply(years_5y, eci_function, df=fiveyear)
fiveyear<-bind_rows(fiveyear)

#Join values to main dataset
fiveyear<-fiveyear%>%left_join(pci_5y, by=c("year","cpc_level12"="cpc"))
fiveyear<-fiveyear%>%left_join(fiveyear, by=c("year","origin"="country"))

#Investigate normality of ECI and PCI values
fiveyear%>%filter(year=="1995-1999")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 1995-1999")
fiveyear%>%filter(year=="2000-2004")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 2000-2004")
fiveyear%>%filter(year=="2005-2009")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 2005-2009")
fiveyear%>%filter(year=="2005-2009")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 2005-2009")
fiveyear%>%filter(year=="2010-2014")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 2010-2014")
fiveyear%>%filter(year=="2015-2019")%>%
    ggplot(aes(eci))+geom_histogram()+ggtitle("ECI 2015-2019")

#Investigate top15 rankings
generalstats<-fiveyear%>%group_by(year)%>%
  summarize(avg_ipfcpc_total=sum(ipf_count)/n_distinct(cpc_level12),total_count=sum(ipf_count))

lead9599<- fiveyear%>%filter(year=="1995-1999")%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(ipf_count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="1995-1999")
lead9599<-lead9599%>%left_join(generalstats, by=c("year"))%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))
            
lead0004<- fiveyear%>%filter(year=="2000-2004")%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(ipf_count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="2000-2004")
lead0004<-lead0004%>%left_join(generalstats, by=c("year"))%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))

lead0509<- fiveyear%>%filter(year=="2005-2009")%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(ipf_count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="2005-2009")
lead0509<-lead0509%>%left_join(generalstats, by=c("year"))%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))

lead1014<- fiveyear%>%filter(year=="2010-2014")%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(ipf_count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="2010-2014")
lead1014<-lead1014%>%left_join(generalstats, by=c("year"))%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))

lead1519<- fiveyear%>%filter(year=="2015-2019")%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(ipf_count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="2015-2019")
lead1519<-lead1519%>%left_join(generalstats, by=c("year"))%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))


#Look at PCI
pci_5y<-pci_5y%>%
  left_join(cpc_text, by=c("cpc"="CPC"))

#Investigate top15 rankings
lead_pci9599<-pci_5y%>%filter(year=="1995-1999")%>%
  select(1,2,5)%>%arrange(desc(pci))
lead_pci0004<-pci_5y%>%filter(year=="2000-2004")%>%
  select(1,2,5)%>%arrange(desc(pci))
lead_pci0509<-pci_5y%>%filter(year=="2005-2009")%>%
  select(1,2,5)%>%arrange(desc(pci))
lead_pci1014<-pci_5y%>%filter(year=="2010-2014")%>%
  select(1,2,5)%>%arrange(desc(pci))
lead_pci1519<-pci_5y%>%filter(year=="2015-2019")%>%
  select(1,2,5)%>%arrange(desc(pci))
```

11. ECI calculation over entire period
```{r}
productspace<-pat_count_country_reduced%>%
  group_by(origin, cpc_level12)%>%
  summarise(count=sum(ipf_count))

 balassa_mat<-balassa_index(
        data = productspace,
        country = "origin",
        product = "cpc_level12",
        value = "count"
      )
  com<-complexity_measures(balassa_mat,
                      method = "eigenvalues")
  eci<-as.data.frame(com$complexity_index_country)
  eci<-setDT(eci,keep.rownames = TRUE)
  eci<-eci%>%rename(country=rn)
  eci<-eci%>%rename(eci="com$complexity_index_country")

productspace<-productspace%>%left_join(eci,by=c("origin"="country"))
    
productspace<-productspace%>%
  group_by(origin)%>%
  mutate(sum_country=sum(count))%>%
  ungroup()%>%
  group_by(cpc_level12)%>%
  mutate(sum_cpc=sum(count))%>%
  ungroup()%>%
  mutate(total_pat=sum(count))%>%
  mutate(rta=(count/sum_country)/(sum_cpc/total_pat))

productspace<-productspace%>%mutate(rta1=if_else(rta>1,1,0))

generalstats2<-productspace%>%
  summarize(avg_ipfcpc_total=sum(count)/n_distinct(cpc_level12),total_count=sum(count))%>%mutate(year="all_periods")

eci_fullperiod<- productspace%>%group_by(origin)%>%
  summarise(eci=mean(eci),ipf_count=sum(count),cpc=n_distinct(cpc_level12), no_rta=sum(rta1))%>%
  mutate(avg_ipf_cpc__country=ipf_count/cpc)%>%mutate(year="all_periods")

eci_fullperiod<-eci_fullperiod%>%bind_cols(generalstats2)
eci_fullperiod<-eci_fullperiod%>%
  mutate(ipf_needed_for_rca=(avg_ipfcpc_total/total_count)*ipf_count)%>%
  mutate(world_share=ipf_count/total_count)%>%
  arrange(desc(eci))

eci_fullperiod%>%filter(world_share>0.01)%>%
  ggplot(aes(x= reorder(origin, -world_share), y=world_share))+geom_bar(stat = "identity")+
  ggtitle("Share of Y02 IPF over full period 1995-2019")+
  geom_text(aes(label=round(world_share,digits=2)),vjust=-0.4)
#7 countries make up 81% of all IPF counts


summary(eci$eci)
sd(eci$eci)
hist(eci$eci)
eci%>%filter(eci>-3)%>%
  ggplot(aes(x=eci))+geom_histogram()+ggtitle("ECI Distribution-ECI calculated over whole period")

```

11. Add the PCI over the full period
```{r}
pci<-as.data.frame(com$complexity_index_product)
pci<-setDT(pci,keep.rownames = TRUE)
pci<-pci%>%rename(cpc=rn)
pci<-pci%>%rename(pci="com$complexity_index_product")

productspace<-productspace%>%left_join(pci,by=c("cpc_level12"="cpc"))

pci_fullperiod<-productspace%>%group_by(cpc_level12)%>%
  summarize(count=sum(count))%>%
  ungroup%>%
  mutate(share=count/sum(count))
pci_fullperiod<-pci_fullperiod%>%left_join(cpc_text, by=c("cpc_level12"="CPC"))

pci_fullperiod%>%filter(share>0.03)%>%
  ggplot(aes(x= reorder(short_name, -share), y=share))+geom_bar(stat = "identity")+
  ggtitle("Share of Y02 CPC codes (level 1/2) of total IPF over full period 1995-2019")+
  geom_text(aes(label=round(share,digits=2)),vjust=-0.4)

#7/168 (~4%) make up 47% of all IPF applications
#Batteries,
```



11. make the product space (0.2 tolerance)
```{r}
library(igraph)
library(ggplot2)
library(ggraph)
#One dataset for the full period 
description<-productspace%>%distinct(cpc_level12)
description<-description%>%left_join(cpc_text, by=c("cpc_level12"="CPC"))%>%arrange(cpc_level12)
bi <- balassa_index(
  data = productspace,
  country = "origin",
  product = "cpc_level12",
  value = "count"
)

com_fit <- complexity_measures(balassa_index = bi)

pro <- proximity(
  balassa_index = bi
)

net <- projections(
  proximity_country = pro$proximity_country,
  proximity_product = pro$proximity_product,
  tolerance = 0.2
)

aggregated_products <- productspace %>%
  group_by(cpc_level12) %>%
  summarise(count = sum(count))

aggregated_products <- setNames(aggregated_products$count, aggregated_products$cpc_level12)
V(net$network_product)$size <- aggregated_products[match(V(net$network_product)$name, names(aggregated_products))]


V(net$network_product)$name2 <- description$area
set.seed(200100)
#New product space with colored nodes

colored_space<-ggraph(net$network_product, layout = "kk") +
  # geom_edge_link(aes(edge_width = weight), edge_colour = "#a8a8a8") +
  geom_edge_link(edge_colour = "#a8a8a8") +
  geom_node_point(aes(size = size,color = factor(unlist(name2))), show.legend= T)+
   #geom_node_text(aes(label = name2), size = 2, vjust = 2.2) +
  ggtitle("Product Space with Main CPC Sectors, proximity tolerance=0.2")+
  theme_void()+
   theme(legend.position = "left")+
  labs(col="Area")+
  scale_color_manual(values=c("adaptation"="#82858A","mitigation"="#FFC100","CCS"="#00FFF3", "ICT"="#FF00F0", "energy"="#00FF3E", "production"="#FF000C", "transport"="#030303", "water"="#035DFA"))+
  geom_node_text(aes(label = name), size=2.5, vjust = 2.2)+
  labs(size="Number of Intl. Patent Families")

colored_space

#save it with width=2000, height=1242
```






12. make the product space (0.05 tolerance)
```{r}

bi <- balassa_index(
  data = productspace,
  country = "origin",
  product = "cpc_level12",
  value = "count"
)

com_fit <- complexity_measures(balassa_index = bi)

pro <- proximity(
  balassa_index = bi
)

net <- projections(
  proximity_country = pro$proximity_country,
  proximity_product = pro$proximity_product,
  tolerance = 0.05
)

aggregated_products <- productspace %>%
  group_by(cpc_level12) %>%
  summarise(count = sum(count))

aggregated_products <- setNames(aggregated_products$count, aggregated_products$cpc_level12)
V(net$network_product)$size <- aggregated_products[match(V(net$network_product)$name, names(aggregated_products))]
V(net$network_product)$name2 <- description$area

set.seed(200100)

colored_space005<-ggraph(net$network_product, layout = "kk") +
  # geom_edge_link(aes(edge_width = weight), edge_colour = "#a8a8a8") +
  geom_edge_link(edge_colour = "#a8a8a8") +
  geom_node_point(aes(size = size,color = factor(unlist(name2))), show.legend= T)+
   #geom_node_text(aes(label = name2), size = 2, vjust = 2.2) +
  ggtitle("Product Space with Main CPC Sectors, proximity tolerance =0.05")+
  theme_void()+
   theme(legend.position = "left")+
  labs(col="Area")+
  scale_color_manual(values=c("adaptation"="#82858A","mitigation"="#FFC100","CCS"="#00FFF3", "ICT"="#FF00F0", "energy"="#00FF3E", "production"="#FF000C", "transport"="#030303", "water"="#035DFA"))+
  geom_node_text(aes(label = name), size=2.5, vjust = 2.2)+
  labs(size="Number of Intl. Patent Families")

colored_space005

```
